wined3d: Directly upload converted formats in surface_upload_from_surface
This way we don't have to map destination resources in update_texture or
update_surface, which are usually not mappable. It also gets rid of the
call circle between surface_upload_from_surface and wined3d_surface_blt.

The drawback is that we can't properly map a dynamic default pool
surface with a converted format that has been updated with
update_surface or update_texture. In the remote event that an
application needs this, my preferred solution is to add support for
converting converted formats back.