wined3d: Use a critical section to pause the main thread.
This isn't exactly a by the books solution. Critical sections
generally aren't suitable for producer-consumer problems.

The solution works because I am using the extra LONG field to make sure
the worker thread doesn't release and then immediately re-enter the
critical section without waking up the main thread. The advantage of
this solution is that it doesn't require the server, nor does it require
difficult to port POSIX synchronization primitives.

This setup does not work in the other direction (for putting the worker
thread to sleep if no commands are available) because more than one
thread may submit commands.